requirements.txt
torchaudio
pytorch
torchvision
scikit-learn
numpy
pathlib
